Implement sync of form data with cache
- Now the common fields listen for the cache value. If the emitter emits
a value with the name of the field, they will update their value.
- Save the form data in the cache after a wizard generation. THe ideal solution
is to save the data after each field changes its value, but we have issues
with the sync of the events.

NOTE: The user should have a config for this. They might not want to load
previous data.